+++ src/acp/java/com/ancientmc/acp/task/step/function/MakeReobfSrg.java
@@ -24,30 +24,33 @@ public class MakeReobfSrg extends Step {
     @Override
     public void exec(Logger logger, boolean condition) {
         super.exec(logger, condition);
-        File temp = project.file(Paths.DIR_TEMP + "temp.srg");
 
-        try {
-            IMappingFile.load(input).reverse().write(temp.toPath(), IMappingFile.Format.SRG, false);
+        if (condition) {
+            File temp = project.file(Paths.DIR_TEMP + "temp.srg");
 
-            BufferedWriter writer = new BufferedWriter(new FileWriter(output));
-            List<String> lines = Files.readAllLines(temp.toPath());
+            try {
+                IMappingFile.load(input).reverse().write(temp.toPath(), IMappingFile.Format.SRG, false);
 
-            /*
-             Adds lines to strip the package of any straggling classes with the "net/minecraft/src" or the "com/mojang/minecraft/src/" packages.
-             Since the vanilla classes are already accounted for in the SRG, by process of elimination this leaves mod classes who get put
-             into the /src/ path.
-             */
-            writer.write("PK: net/minecraft/src .\n");
-            writer.write("PK: com/mojang/minecraft/src .\n");
+                BufferedWriter writer = new BufferedWriter(new FileWriter(output));
+                List<String> lines = Files.readAllLines(temp.toPath());
 
-            for(String line : lines) {
-                writer.write(line + "\n");
+                /*
+                Adds lines to strip the package of any straggling classes with the "net/minecraft/src" or the "com/mojang/minecraft/src/" packages.
+                Since the vanilla classes are already accounted for in the SRG, by process of elimination this leaves mod classes who get put
+                into the /src/ path.
+                */
+                writer.write("PK: net/minecraft/src .\n");
+                writer.write("PK: com/mojang/minecraft/src .\n");
+
+                for(String line : lines) {
+                    writer.write(line + "\n");
+                }
+
+                writer.close();
+                FileUtils.forceDelete(temp);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
             }
-
-            writer.close();
-            FileUtils.forceDelete(temp);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
         }
     }
